requirements.txt
opencv-python>=4.8.0
flask>=3.0.0
numpy>=1.26.0
pillow>=10.0.0
werkzeug>=3.0.0
click>=8.1.0
python-dotenv>=1.0.0 